back-end/requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
fastapi==0.74.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
inflection==0.5.1
numpy==1.22.2
pandas==1.4.1
pycparser==2.21
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
tabulate==0.8.9
typing_extensions==4.1.1
uvicorn==0.17.5
